etl/requirements.txt
python-slugify==8.0.1
openai